+++ src/main/kotlin/io/openfuture/chain/service/DefaultBlockService.kt
@@ -1,6 +1,5 @@
 package io.openfuture.chain.service
 
-import io.openfuture.chain.domain.block.BlockRequest
 import io.openfuture.chain.entity.Block
 import io.openfuture.chain.exception.NotFoundException
 import io.openfuture.chain.repository.BlockRepository
@@ -8,24 +7,21 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
-class DefaultBlockService (
-        private val blockRepository: BlockRepository
+class DefaultBlockService(
+    private val blockRepository: BlockRepository
 ) : BlockService {
 
     @Transactional(readOnly = true)
-    override fun get(id: Int): Block = blockRepository.getOne(id)
-        ?: throw NotFoundException("Not found id $id")
-
+    override fun get(hash: String): Block = blockRepository.findByHash(hash)
+        ?: throw NotFoundException("Block with hash:$hash not found")
 
     @Transactional(readOnly = true)
-    override fun getAll(): MutableList<Block> =  blockRepository.findAll()
+    override fun getLast(): Block = blockRepository.findFirstByOrderByHeightDesc()
+        ?: throw NotFoundException("Last block not found!")
 
-    override fun getLast(): Block = blockRepository.findFirstByOrderByOrderNumberDesc()
-        ?: throw NotFoundException("Last block not exist!")
-
-
-    override fun save(request: BlockRequest): Block {
-        return blockRepository.save(Block.of(request))
+    @Transactional(readOnly = true)
+    override fun getLastGenesis(): Block {
+        TODO("not implemented")
     }
 
 }