+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - RPC requests: validateAddress
 - Sign data
 - Verify signed data
+- Define epoch duration in properties
+- Get current epoch height
+- Check is genesis block needed
 
 ### Changed
 - Default account info is added to the doGenerateMaster end-point response